Refactor tests to reflect issue
