+++ core/src/test/scala/ai/lum/odinson/metadata/TestJoinQueries.scala
@@ -1,0 +1,36 @@
+package ai.lum.odinson.metadata
+
+import ai.lum.odinson.test.utils.OdinsonTest
+
+class TestJoinQueries extends OdinsonTest {
+
+  "Odinson" should "match simple patterns with open range quantifiers and a metadata filter" in {
+
+    // create in-memory engine w/ a single doc w/ a single sentence
+    val ee = mkExtractorEngine("step-bros")
+    // "John", "C.", "Reilly", "played" ...
+    val pattern = "[lemma=play] >nsubj [tag=NNP]{,3}"
+    val mf = "doc_id == 'step-bros'"
+    val oq = ee.mkFilteredQuery(query = pattern, metadataFilter = mf)
+
+    val res = ee.query(oq)
+
+    numMatches(res) should be (1)
+    existsMatchWithSpan(odinResults = res, doc = 0, start = 0, end = 3) should be (true)
+  }
+
+  it should "match simple patterns with exact range quantifiers (>=2) and a metadata filter" in {
+
+    // create in-memory engine w/ a single doc w/ a single sentence
+    val ee = mkExtractorEngine("step-bros")
+    // "John", "C.", "Reilly", "played" ...
+    val pattern = "[lemma=play] >nsubj [tag=NNP]{3}"
+    val mf = "doc_id == 'step-bros'"
+    val oq = ee.mkFilteredQuery(query = pattern, metadataFilter = mf)
+
+    val res = ee.query(oq)
+
+    numMatches(res) should be (1)
+    existsMatchWithSpan(odinResults = res, doc = 0, start = 0, end = 3) should be (true)
+  }
+}
